# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task ",
        "description": "Game keys marketplace",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payment": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Get list of Payments",
                "description": "Returns list of payments",
                "operationId": "318a8e6d1447a7cbd66237ccd6660678",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Payment"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "payments"
                ],
                "summary": "New Payment",
                "operationId": "8bd089442f313f34b467e2625ff83655",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "good_id": {
                                        "type": "string"
                                    },
                                    "credit_card": {
                                        "type": "string"
                                    },
                                    "buyer_email": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "data": {
                                            "properties": {
                                                "good_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The good id field is required."
                                                    }
                                                },
                                                "credit_card": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The credit card field is required."
                                                    }
                                                },
                                                "buyer_email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The buyer email field is required."
                                                    }
                                                },
                                                "price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The price field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payment/{id}": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Get Payment by Id",
                "description": "Return payment by id",
                "operationId": "5ec820dc5a59156a714c50a583e10016",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id to return",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Payment Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Payment": {
                "title": "Payment",
                "description": "Payment Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer"
                    },
                    "good_id": {
                        "title": "good id",
                        "description": "good id",
                        "type": "string"
                    },
                    "credit_card": {
                        "title": "credit_card",
                        "description": "credit_card",
                        "type": "string"
                    },
                    "buyer_email": {
                        "title": "buyer_email",
                        "description": "buyer_email",
                        "type": "string"
                    },
                    "price": {
                        "title": "price",
                        "description": "price",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}